Download titles, URLs, durations of a Playlist
You just need to open the YouTube Playlist that you need to work on. Then go to the inspect option by using the right mouse. Then you need to go to the Chrome Console and paste the code. Then a CSV file will be downloaded with all the video titles, links, duration and duration in seconds.